requirements.txt
absl-py @ file:///croot/absl-py_1714140470852/work
accelerate==1.1.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.2.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
DataProperty==1.1.0
datasets==3.3.2
diffusers==0.24.0
dill==0.3.7
docker-pycreds==0.4.0
docstring_parser==0.16
evaluate==0.4.3
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2023.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
grpcio @ file:///croot/grpc-split_1716834572340/work
huggingface-hub==0.25.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.6
joblib @ file:///croot/joblib_1718217211762/work
jsonlines==4.0.0
kiwisolver==1.4.7
littleutils==0.2.4
# Editable install with no version control (lm_eval==0.4.8)
-e /home/cgq/rlds/lm-evaluation-harness
lxml==5.4.0
Markdown @ file:///croot/markdown_1671541909495/work
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.7.2
mbstrdecoder==1.1.4
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
more-itertools==10.7.0
multidict==6.1.0
multiprocess==0.70.15
nltk==3.9.1
numexpr==2.10.2
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
ogb==1.3.6
opencv-python==4.10.0.84
outdated==0.2.2
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.3
peft==0.15.2
pillow==11.0.0
platformdirs==4.3.6
portalocker==3.1.1
propcache==0.3.0
protobuf==4.25.3
psutil==6.1.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pybind11==2.13.6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.0.9
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge_score==0.1.2
sacrebleu==2.5.1
safetensors==0.4.5
scikit-learn==1.4.2
scikit-opt==0.6.6
scipy==1.10.1
sentry-sdk==2.22.0
setproctitle==1.3.5
shtab==1.7.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
sqlitedict==2.1.0
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tensorboard @ file:///croot/tensorboard_1723042039441/work/tensorboard-2.17.0-py3-none-any.whl#sha256=859a499a9b1fb68a058858964486627100b71fcb21646861c61d31846a6478fb
tensorboard_data_server @ file:///croot/tensorboard-data-server_1724172556187/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl#sha256=4428d660733184391dab7b6b6a953467b614b984df0cc1baf3d8c5d3d4ecf29d
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
tokenizers==0.21.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.67.0
tqdm-multiprocess==0.0.11
transformers==4.47.0
trl==0.9.4
typeguard==4.4.2
typepy==1.3.4
typing_extensions==4.12.2
tyro==0.9.17
tzdata==2024.2
urllib3==2.2.3
wandb==0.19.8
wcwidth==0.2.13
Werkzeug @ file:///croot/werkzeug_1731005776056/work
wilds==2.0.0
word2number==1.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
